requirements.txt
python-gnupg
pyzmq
prompt_toolkit